more bugfixes :D

+++ VaultLib.Core/Utils/VLT64Hasher.cs
@@ -13,7 +13,7 @@ namespace VaultLib.Core.Utils
         {
             if (string.IsNullOrEmpty(str) && (str == null || returnZeroForEmpty))
                 return 0;
-            if (str.StartsWith("0x") && uint.TryParse(str.Substring(2), NumberStyles.AllowHexSpecifier, CultureInfo.CurrentCulture, out uint u))
+            if (str.StartsWith("0x") && ulong.TryParse(str.Substring(2), NumberStyles.AllowHexSpecifier, CultureInfo.CurrentCulture, out ulong u))
                 return u;
             byte[] data = Encoding.ASCII.GetBytes(str);
             var bytesProcessed = (uint)str.Length;
